chore: updates as per PR review/feedbacks
